[PD] Update common backend and refactor mooncake backend
